POJO中float改成double

+++ src/main/java/com/fudanscetool/springboot/pojo/HistoryStage1.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class HistoryStage1 {
     private String projectID;
 
-    private float size;
+    private double size;
 }
